kmeans prediction added
updated server.R and ui.R to allow for a prediction radio button.  If selected, instead of having save write results, it will repredict the data and write the repredicted results.  And, if a user should open a new file after creating a kmeans solution, the prediction will occur again (as long as the radio button is selected.  Ideally, we will eventually have an "open new data file to be predicted" tab, but this should work for now.