Added SOM Prediction to Save tab
As clunky as it seems, this working version of server.R and ui.R contain the 'predict' function which predicts new observations from a SOM solution.  So, if the user first selects the radio button for "Predict Kmeans" and saves, they get a spreadsheet with the Kmeans cluster.  Then, if they go to the SOM tab and train the SOM, they can return to kmeans tab and check the "Predict SOM" radio button instead.  When saving this time, they get a column in their spreadsheet with the SOM neurons most closely mapped.  (Then, outside of R, they can compare these two spreadsheets to see how the Neurons overlay with the Clusters.)